pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lettaplus"
version = "0.1.0"
description = "Self-evolving agentic AI architecture with Dynamic Capabilities Framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
skill = "skill_cli.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["skill_cli*", "dcf_mcp*", "stub_mcp*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
